+++ schemars_derive/src/attr/validation.rs
@@ -1,21 +1,20 @@
-use super::{expr_as_lit_str, get_meta_items, parse_lit_str};
 use proc_macro2::TokenStream;
-use quote::ToTokens;
-use serde_derive_internals::Ctxt;
-use syn::{
-    parse::Parser, punctuated::Punctuated, Expr, ExprLit, ExprPath, Lit, Meta, MetaList,
-    MetaNameValue, Path,
+use syn::{Expr, Meta};
+
+use crate::idents::SCHEMA;
+
+use super::{
+    parse_meta::{
+        parse_contains, parse_length_or_range, parse_nested_meta, parse_schemars_regex,
+        parse_validate_regex, require_path_only, LengthOrRange,
+    },
+    AttrCtxt,
 };
 
-pub(crate) static VALIDATION_KEYWORDS: &[&str] = &[
-    "range", "regex", "contains", "email", "phone", "url", "length", "required", "inner",
-];
-
 #[derive(Debug, Clone, Copy, PartialEq)]
-enum Format {
+pub enum Format {
     Email,
     Uri,
-    Phone,
 }
 
 impl Format {
@@ -23,7 +22,6 @@ impl Format {
         match self {
             Format::Email => "email",
             Format::Uri => "url",
-            Format::Phone => "phone",
         }
     }
 
@@ -31,420 +29,167 @@ impl Format {
         match self {
             Format::Email => "email",
             Format::Uri => "uri",
-            Format::Phone => "phone",
+        }
+    }
+
+    fn from_attr_str(s: &str) -> Self {
+        match s {
+            "email" => Format::Email,
+            "url" => Format::Uri,
+            _ => panic!("Invalid format attr string `{s}`. This is a bug in schemars, please raise an issue!"),
         }
     }
 }
 
 #[derive(Debug, Default)]
 pub struct ValidationAttrs {
-    length_min: Option<Expr>,
-    length_max: Option<Expr>,
-    length_equal: Option<Expr>,
-    range_min: Option<Expr>,
-    range_max: Option<Expr>,
-    regex: Option<Expr>,
-    contains: Option<String>,
-    required: bool,
-    format: Option<Format>,
-    inner: Option<Box<ValidationAttrs>>,
+    pub length: Option<LengthOrRange>,
+    pub range: Option<LengthOrRange>,
+    pub regex: Option<Expr>,
+    pub contains: Option<Expr>,
+    pub required: bool,
+    pub format: Option<Format>,
+    pub inner: Option<Box<ValidationAttrs>>,
 }
 
 impl ValidationAttrs {
-    pub fn new(attrs: &[syn::Attribute], errors: &Ctxt) -> Self {
-        let schemars_items = get_meta_items(attrs, "schemars", errors, false);
-        let validate_items = get_meta_items(attrs, "validate", errors, true);
-
-        ValidationAttrs::default()
-            .populate(schemars_items, "schemars", false, errors)
-            .populate(validate_items, "validate", true, errors)
+    pub fn add_mutators(&self, mutators: &mut Vec<TokenStream>) {
+        self.add_mutators2(mutators, &quote!(&mut #SCHEMA));
     }
 
-    pub fn required(&self) -> bool {
-        self.required
-    }
-
-    fn populate(
-        mut self,
-        meta_items: Vec<Meta>,
-        attr_type: &'static str,
-        ignore_errors: bool,
-        errors: &Ctxt,
-    ) -> Self {
-        let duplicate_error = |path: &Path| {
-            if !ignore_errors {
-                let msg = format!(
-                    "duplicate schemars attribute `{}`",
-                    path.get_ident().unwrap()
-                );
-                errors.error_spanned_by(path, msg)
-            }
-        };
-        let mutual_exclusive_error = |path: &Path, other: &str| {
-            if !ignore_errors {
-                let msg = format!(
-                    "schemars attribute cannot contain both `{}` and `{}`",
-                    path.get_ident().unwrap(),
-                    other,
-                );
-                errors.error_spanned_by(path, msg)
-            }
-        };
-        let duplicate_format_error = |existing: Format, new: Format, path: &syn::Path| {
-            if !ignore_errors {
-                let msg = if existing == new {
-                    format!("duplicate schemars attribute `{}`", existing.attr_str())
-                } else {
-                    format!(
-                        "schemars attribute cannot contain both `{}` and `{}`",
-                        existing.attr_str(),
-                        new.attr_str(),
-                    )
-                };
-                errors.error_spanned_by(path, msg)
-            }
-        };
-        let parse_nested_meta = |meta_list: MetaList| {
-            let parser = Punctuated::<syn::Meta, Token![,]>::parse_terminated;
-            match parser.parse2(meta_list.tokens) {
-                Ok(p) => p,
-                Err(e) => {
-                    if !ignore_errors {
-                        errors.syn_error(e);
-                    }
-                    Default::default()
-                }
-            }
-        };
-
-        for meta_item in meta_items {
-            match meta_item {
-                Meta::List(meta_list) if meta_list.path.is_ident("length") => {
-                    for nested in parse_nested_meta(meta_list) {
-                        match nested {
-                            Meta::NameValue(nv) if nv.path.is_ident("min") => {
-                                if self.length_min.is_some() {
-                                    duplicate_error(&nv.path)
-                                } else if self.length_equal.is_some() {
-                                    mutual_exclusive_error(&nv.path, "equal")
-                                } else {
-                                    self.length_min = str_or_num_to_expr(errors, "min", nv.value);
-                                }
-                            }
-                            Meta::NameValue(nv) if nv.path.is_ident("max") => {
-                                if self.length_max.is_some() {
-                                    duplicate_error(&nv.path)
-                                } else if self.length_equal.is_some() {
-                                    mutual_exclusive_error(&nv.path, "equal")
-                                } else {
-                                    self.length_max = str_or_num_to_expr(errors, "max", nv.value);
-                                }
-                            }
-                            Meta::NameValue(nv) if nv.path.is_ident("equal") => {
-                                if self.length_equal.is_some() {
-                                    duplicate_error(&nv.path)
-                                } else if self.length_min.is_some() {
-                                    mutual_exclusive_error(&nv.path, "min")
-                                } else if self.length_max.is_some() {
-                                    mutual_exclusive_error(&nv.path, "max")
-                                } else {
-                                    self.length_equal =
-                                        str_or_num_to_expr(errors, "equal", nv.value);
-                                }
-                            }
-                            meta => {
-                                if !ignore_errors {
-                                    errors.error_spanned_by(
-                                        meta,
-                                        "unknown item in schemars length attribute",
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-
-                Meta::List(meta_list) if meta_list.path.is_ident("range") => {
-                    for nested in parse_nested_meta(meta_list) {
-                        match nested {
-                            Meta::NameValue(nv) if nv.path.is_ident("min") => {
-                                if self.range_min.is_some() {
-                                    duplicate_error(&nv.path)
-                                } else {
-                                    self.range_min = str_or_num_to_expr(errors, "min", nv.value);
-                                }
-                            }
-                            Meta::NameValue(nv) if nv.path.is_ident("max") => {
-                                if self.range_max.is_some() {
-                                    duplicate_error(&nv.path)
-                                } else {
-                                    self.range_max = str_or_num_to_expr(errors, "max", nv.value);
-                                }
-                            }
-                            meta => {
-                                if !ignore_errors {
-                                    errors.error_spanned_by(
-                                        meta,
-                                        "unknown item in schemars range attribute",
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-
-                Meta::Path(m) if m.is_ident("required") || m.is_ident("required_nested") => {
-                    self.required = true;
-                }
-
-                Meta::Path(p) if p.is_ident(Format::Email.attr_str()) => match self.format {
-                    Some(f) => duplicate_format_error(f, Format::Email, &p),
-                    None => self.format = Some(Format::Email),
-                },
-                Meta::Path(p) if p.is_ident(Format::Uri.attr_str()) => match self.format {
-                    Some(f) => duplicate_format_error(f, Format::Uri, &p),
-                    None => self.format = Some(Format::Uri),
-                },
-                Meta::Path(p) if p.is_ident(Format::Phone.attr_str()) => match self.format {
-                    Some(f) => duplicate_format_error(f, Format::Phone, &p),
-                    None => self.format = Some(Format::Phone),
-                },
-
-                Meta::NameValue(nv) if nv.path.is_ident("regex") => {
-                    match (&self.regex, &self.contains) {
-                        (Some(_), _) => duplicate_error(&nv.path),
-                        (None, Some(_)) => mutual_exclusive_error(&nv.path, "contains"),
-                        (None, None) => {
-                            self.regex = parse_regex_expr(errors, nv.value);
-                        }
-                    }
-                }
-
-                Meta::List(meta_list) if meta_list.path.is_ident("regex") => {
-                    match (&self.regex, &self.contains) {
-                        (Some(_), _) => duplicate_error(&meta_list.path),
-                        (None, Some(_)) => mutual_exclusive_error(&meta_list.path, "contains"),
-                        (None, None) => {
-                            for x in parse_nested_meta(meta_list) {
-                                match x {
-                                    Meta::NameValue(MetaNameValue { path, value, .. })
-                                        if path.is_ident("path") =>
-                                    {
-                                        self.regex = parse_regex_expr(errors, value);
-                                    }
-                                    Meta::NameValue(MetaNameValue { path, value, .. })
-                                        if path.is_ident("pattern") =>
-                                    {
-                                        self.regex =
-                                            expr_as_lit_str(errors, attr_type, "pattern", &value)
-                                                .ok()
-                                                .map(|litstr| {
-                                                    Expr::Lit(ExprLit {
-                                                        attrs: Vec::new(),
-                                                        lit: Lit::Str(litstr.clone()),
-                                                    })
-                                                })
-                                    }
-                                    meta => {
-                                        if !ignore_errors {
-                                            errors.error_spanned_by(
-                                                meta,
-                                                "unknown item in schemars regex attribute",
-                                            );
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                Meta::NameValue(MetaNameValue { path, value, .. }) if path.is_ident("contains") => {
-                    match (&self.contains, &self.regex) {
-                        (Some(_), _) => duplicate_error(&path),
-                        (None, Some(_)) => mutual_exclusive_error(&path, "regex"),
-                        (None, None) => {
-                            self.contains = expr_as_lit_str(errors, attr_type, "contains", &value)
-                                .map(|litstr| litstr.value())
-                                .ok()
-                        }
-                    }
-                }
-
-                Meta::List(meta_list) if meta_list.path.is_ident("contains") => {
-                    match (&self.contains, &self.regex) {
-                        (Some(_), _) => duplicate_error(&meta_list.path),
-                        (None, Some(_)) => mutual_exclusive_error(&meta_list.path, "regex"),
-                        (None, None) => {
-                            for x in parse_nested_meta(meta_list) {
-                                match x {
-                                    Meta::NameValue(MetaNameValue { path, value, .. })
-                                        if path.is_ident("pattern") =>
-                                    {
-                                        self.contains =
-                                            expr_as_lit_str(errors, attr_type, "contains", &value)
-                                                .ok()
-                                                .map(|litstr| litstr.value())
-                                    }
-                                    meta => {
-                                        if !ignore_errors {
-                                            errors.error_spanned_by(
-                                                meta,
-                                                "unknown item in schemars contains attribute",
-                                            );
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                Meta::List(meta_list) if meta_list.path.is_ident("inner") => match self.inner {
-                    Some(_) => duplicate_error(&meta_list.path),
-                    None => {
-                        let inner_attrs = ValidationAttrs::default().populate(
-                            parse_nested_meta(meta_list).into_iter().collect(),
-                            attr_type,
-                            ignore_errors,
-                            errors,
-                        );
-                        self.inner = Some(Box::new(inner_attrs));
-                    }
-                },
-
-                _ if ignore_errors => {}
-
-                _ => {
-                    if let Some(ident) = meta_item.path().get_ident() {
-                        if VALIDATION_KEYWORDS.iter().any(|k| ident == k) {
-                            errors.error_spanned_by(
-                                &meta_item,
-                                format!("could not parse `{ident}` item in schemars attribute"),
-                            );
-                        }
-                    }
-                }
-            }
-        }
-        self
-    }
-
-    pub fn apply_to_schema(&self, schema_expr: &mut TokenStream) {
-        let setters = self.make_setters(quote!(&mut schema));
-        if !setters.is_empty() {
-            *schema_expr = quote!({
-                let mut schema = #schema_expr;
-                #(#setters)*
-                schema
-            });
-        }
-    }
-
-    fn make_setters(&self, mut_schema: impl ToTokens) -> Vec<TokenStream> {
-        let mut result = Vec::new();
-
-        if let Some(length_min) = self.length_min.as_ref().or(self.length_equal.as_ref()) {
-            result.push(quote! {
-                schemars::_private::insert_validation_property(#mut_schema, "string", "minLength", #length_min);
-            });
-            result.push(quote! {
-                schemars::_private::insert_validation_property(#mut_schema, "array", "minItems", #length_min);
-            });
+    fn add_mutators2(&self, mutators: &mut Vec<TokenStream>, mut_ref_schema: &TokenStream) {
+        if let Some(length) = &self.length {
+            Self::add_length_or_range(length, mutators, "string", "Length", mut_ref_schema);
+            Self::add_length_or_range(length, mutators, "array", "Items", mut_ref_schema);
         }
 
-        if let Some(length_max) = self.length_max.as_ref().or(self.length_equal.as_ref()) {
-            result.push(quote! {
-                schemars::_private::insert_validation_property(#mut_schema, "string", "maxLength", #length_max);
-            });
-            result.push(quote! {
-                schemars::_private::insert_validation_property(#mut_schema, "array", "maxItems", #length_max);
-            });
-        }
-
-        if let Some(range_min) = &self.range_min {
-            result.push(quote! {
-                schemars::_private::insert_validation_property(#mut_schema, "number", "minimum", #range_min);
-            });
-        }
-
-        if let Some(range_max) = &self.range_max {
-            result.push(quote! {
-                schemars::_private::insert_validation_property(#mut_schema, "number", "maximum", #range_max);
-            });
+        if let Some(range) = &self.range {
+            Self::add_length_or_range(range, mutators, "number", "imum", mut_ref_schema);
         }
 
         if let Some(regex) = &self.regex {
-            result.push(quote! {
-                schemars::_private::insert_validation_property(#mut_schema, "string", "pattern", #regex);
+            mutators.push(quote! {
+                schemars::_private::insert_validation_property(#mut_ref_schema, "string", "pattern", #regex.to_string());
             });
         }
 
         if let Some(contains) = &self.contains {
-            result.push(quote! {
-                schemars::_private::append_required(#mut_schema, #contains);
+            mutators.push(quote! {
+                schemars::_private::must_contain(#mut_ref_schema, #contains.to_string());
             });
-
-            if self.regex.is_none() {
-                let pattern = crate::regex_syntax::escape(contains);
-                result.push(quote! {
-                    schemars::_private::insert_validation_property(#mut_schema, "string", "pattern", #pattern);
-                });
-            }
         }
 
         if let Some(format) = &self.format {
             let f = format.schema_str();
-            result.push(quote! {
-                schema.ensure_object().insert("format".into(), #f.into());
+            mutators.push(quote! {
+                (#mut_ref_schema).ensure_object().insert("format".into(), #f.into());
             })
         };
 
         if let Some(inner) = &self.inner {
-            let inner_setters = inner.make_setters(quote!(schema));
-            if !inner_setters.is_empty() {
-                result.push(quote! {
-                    schemars::_private::apply_inner_validation(#mut_schema, |schema| { #(#inner_setters)* });
+            let mut inner_mutators = Vec::new();
+            inner.add_mutators2(&mut inner_mutators, &quote!(inner_schema));
+
+            if !inner_mutators.is_empty() {
+                mutators.push(quote! {
+                    schemars::_private::apply_inner_validation(#mut_ref_schema, |inner_schema| { #(#inner_mutators)* });
                 })
             }
         }
-
-        result
     }
-}
 
-fn str_or_num_to_expr(cx: &Ctxt, meta_item_name: &str, expr: Expr) -> Option<Expr> {
-    // this odd double-parsing is to make `-10` parsed as an Lit instead of an Expr::Unary
-    let lit: Lit = match syn::parse2(expr.to_token_stream()) {
-        Ok(l) => l,
-        Err(err) => {
-            cx.syn_error(err);
-            return None;
+    fn add_length_or_range(
+        value: &LengthOrRange,
+        mutators: &mut Vec<TokenStream>,
+        required_format: &str,
+        key_suffix: &str,
+        mut_ref_schema: &TokenStream,
+    ) {
+        if let Some(min) = value.min.as_ref().or(value.equal.as_ref()) {
+            let key = format!("min{key_suffix}");
+            mutators.push(quote!{
+                schemars::_private::insert_validation_property(#mut_ref_schema, #required_format, #key, #min);
+            });
         }
-    };
 
-    match lit {
-        Lit::Str(s) => parse_lit_str::<ExprPath>(&s).ok().map(Expr::Path),
-        Lit::Int(_) | Lit::Float(_) => Some(expr),
-        _ => {
-            cx.error_spanned_by(
-                &expr,
-                format!(
-                    "expected `{}` to be a string or number literal, not {:?}",
-                    meta_item_name, &expr
-                ),
-            );
-            None
+        if let Some(max) = value.max.as_ref().or(value.equal.as_ref()) {
+            let key = format!("max{key_suffix}");
+            mutators.push(quote!{
+                schemars::_private::insert_validation_property(#mut_ref_schema, #required_format, #key, #max);
+            });
         }
     }
-}
 
-fn parse_regex_expr(cx: &Ctxt, value: Expr) -> Option<Expr> {
-    match value {
-        Expr::Lit(ExprLit {
-            lit: Lit::Str(litstr),
-            ..
-        }) => parse_lit_str(&litstr).map_err(|e| cx.syn_error(e)).ok(),
-        v => Some(v),
+    pub(super) fn populate(&mut self, schemars_cx: &mut AttrCtxt, validate_cx: &mut AttrCtxt) {
+        self.process_attr(schemars_cx);
+        self.process_attr(validate_cx);
+    }
+
+    fn process_attr(&mut self, cx: &mut AttrCtxt) {
+        cx.parse_meta(|m, n, c| self.process_meta(m, n, c));
+    }
+
+    fn process_meta(&mut self, meta: Meta, meta_name: &str, cx: &AttrCtxt) -> Option<Meta> {
+        match meta_name {
+            "length" => match self.length {
+                Some(_) => cx.duplicate_error(&meta),
+                None => self.length = parse_length_or_range(meta, cx).ok(),
+            },
+
+            "range" => match self.range {
+                Some(_) => cx.duplicate_error(&meta),
+                None => self.range = parse_length_or_range(meta, cx).ok(),
+            },
+
+            "email" | "url" => self.handle_format(meta, meta_name, cx),
+
+            "required" => {
+                if self.required {
+                    cx.duplicate_error(&meta);
+                } else if require_path_only(meta, cx).is_ok() {
+                    self.required = true;
+                }
+            }
+
+            "regex" => match (&self.regex, &self.contains, cx.attr_type) {
+                (Some(_), _, _) => cx.duplicate_error(&meta),
+                (_, Some(_), _) => cx.mutual_exclusive_error(&meta, "contains"),
+                (None, None, "schemars") => self.regex = parse_schemars_regex(meta, cx).ok(),
+                (None, None, "validate") => self.regex = parse_validate_regex(meta, cx).ok(),
+                (None, None, wat) => panic!("Unexpected attr type `{wat}` for regex item. This is a bug in schemars, please raise an issue!"),
+            },
+            "contains" => match (&self.regex, &self.contains) {
+                (Some(_), _) => cx.mutual_exclusive_error(&meta, "regex"),
+                (_, Some(_)) => cx.duplicate_error(&meta),
+                (None, None) => self.contains = parse_contains(meta, cx).ok(),
+            },
+
+            "inner" => {
+                if let Ok(nested_meta) = parse_nested_meta(meta, cx) {
+                    let inner = self
+                        .inner
+                        .get_or_insert_with(|| Box::new(ValidationAttrs::default()));
+                    let mut inner_cx = cx.new_nested_meta(nested_meta.into_iter().collect());
+                    inner.process_attr(&mut inner_cx);
+                }
+            }
+
+            _ => return Some(meta),
+        }
+
+        None
+    }
+
+    fn handle_format(&mut self, meta: Meta, meta_name: &str, cx: &AttrCtxt) {
+        match &self.format {
+            Some(f) if f.attr_str() == meta_name => cx.duplicate_error(&meta),
+            Some(f) => cx.mutual_exclusive_error(&meta, f.attr_str()),
+            None => {
+                // FIXME this is too strict - it may be a MetaList in validator attr (e.g. with message/code items)
+                if require_path_only(meta, cx).is_ok() {
+                    self.format = Some(Format::from_attr_str(meta_name))
+                }
+            }
+        }
     }
 }