+++ schemars_derive/src/lib.rs
@@ -8,11 +8,11 @@ extern crate proc_macro;
 
 mod ast;
 mod attr;
-mod metadata;
-mod regex_syntax;
+mod idents;
 mod schema_exprs;
 
 use ast::*;
+use idents::GENERATOR;
 use proc_macro2::TokenStream;
 use syn::spanned::Spanned;
 
@@ -70,12 +70,12 @@ fn derive_json_schema(mut input: syn::DeriveInput, repr: bool) -> syn::Result<To
                         <#ty as schemars::JsonSchema>::schema_id()
                     }
 
-                    fn json_schema(generator: &mut schemars::SchemaGenerator) -> schemars::Schema {
-                        <#ty as schemars::JsonSchema>::json_schema(generator)
+                    fn json_schema(#GENERATOR: &mut schemars::SchemaGenerator) -> schemars::Schema {
+                        <#ty as schemars::JsonSchema>::json_schema(#GENERATOR)
                     }
 
-                    fn _schemars_private_non_optional_json_schema(generator: &mut schemars::SchemaGenerator) -> schemars::Schema {
-                        <#ty as schemars::JsonSchema>::_schemars_private_non_optional_json_schema(generator)
+                    fn _schemars_private_non_optional_json_schema(#GENERATOR: &mut schemars::SchemaGenerator) -> schemars::Schema {
+                        <#ty as schemars::JsonSchema>::_schemars_private_non_optional_json_schema(#GENERATOR)
                     }
 
                     fn _schemars_private_is_option() -> bool {
@@ -188,7 +188,7 @@ fn derive_json_schema(mut input: syn::DeriveInput, repr: bool) -> syn::Result<To
                     #schema_id
                 }
 
-                fn json_schema(generator: &mut schemars::SchemaGenerator) -> schemars::Schema {
+                fn json_schema(#GENERATOR: &mut schemars::SchemaGenerator) -> schemars::Schema {
                     #schema_expr
                 }
             };